+++ src/main/java/dev/gacbl/logicore/core/CreativeTabModule.java
@@ -7,6 +7,7 @@ import dev.gacbl.logicore.blocks.datacable.DataCableModule;
 import dev.gacbl.logicore.blocks.datacenter.DatacenterModule;
 import dev.gacbl.logicore.blocks.datacenter_port.DatacenterPortModule;
 import dev.gacbl.logicore.blocks.drone_bay.DroneBayModule;
+import dev.gacbl.logicore.blocks.generator.GeneratorModule;
 import dev.gacbl.logicore.blocks.serverrack.ServerRackModule;
 import dev.gacbl.logicore.entity.drone.DroneModule;
 import dev.gacbl.logicore.items.processorunit.ProcessorUnitModule;
@@ -35,6 +36,7 @@ public class CreativeTabModule {
                         output.accept(DatacenterPortModule.DATACENTER_PORT.get());
                         output.accept(DroneBayModule.DRONE_BAY.get());
                         output.accept(DroneModule.DRONE_ITEM.get());
+                        output.accept(GeneratorModule.GENERATOR.get());
                     })
                     .build());