+++ src/main/java/dev/gacbl/logicore/data/ModRecipeProvider.java
@@ -6,6 +6,7 @@ import dev.gacbl.logicore.blocks.datacable.DataCableBlock;
 import dev.gacbl.logicore.blocks.datacenter.DatacenterControllerBlock;
 import dev.gacbl.logicore.blocks.datacenter_port.DatacenterPortBlock;
 import dev.gacbl.logicore.blocks.drone_bay.DroneBayBlock;
+import dev.gacbl.logicore.blocks.generator.GeneratorBlock;
 import dev.gacbl.logicore.blocks.serverrack.ServerRackBlock;
 import dev.gacbl.logicore.entity.drone.DroneItem;
 import dev.gacbl.logicore.items.processorunit.ProcessorUnitItem;
@@ -35,6 +36,7 @@ public class ModRecipeProvider extends RecipeProvider implements IConditionBuild
         DatacenterPortBlock.getRecipe().unlockedBy("has_redstone", has(Items.REDSTONE)).save(recipeOutput);
         DroneBayBlock.getRecipe().unlockedBy("has_redstone", has(Items.REDSTONE)).save(recipeOutput);
         DroneItem.getRecipe().unlockedBy("has_redstone", has(Items.REDSTONE)).save(recipeOutput);
+        GeneratorBlock.getRecipe().unlockedBy("has_redstone", has(Items.REDSTONE)).save(recipeOutput);
 
         /*CompilerRecipeBuilder.of(
                         Ingredient.of(Items.IRON_INGOT), // Input