+++ src/main/java/dev/gacbl/logicore/entity/drone/goals/FollowPlayerGoal.java
@@ -18,7 +18,6 @@ public class FollowPlayerGoal extends Goal {
     @Override
     public boolean canUse() {
         this.owner = drone.getOwner();
-        // Don't follow if need recharge
         return owner != null && !owner.isSpectator() && drone.getEnergy() > 0 && drone.distanceToSqr(owner) > 10.0;
     }
 
@@ -29,9 +28,9 @@ public class FollowPlayerGoal extends Goal {
         drone.getLookControl().setLookAt(owner, 10.0F, (float) drone.getMaxHeadXRot());
 
         // Target position: Behind owner and slightly up
-        double targetX = owner.getX() - (owner.getLookAngle().x * 2);
-        double targetZ = owner.getZ() - (owner.getLookAngle().z * 2);
-        double targetY = owner.getY() + 2.5; // Above head
+        double targetX = owner.getX() - (owner.getLookAngle().x * 1.2);
+        double targetZ = owner.getZ() - (owner.getLookAngle().z * 1.2);
+        double targetY = owner.getY() + 3.5;
 
         drone.getNavigation().moveTo(targetX, targetY, targetZ, 1.0D);
     }